ENH: Add preliminary support for dates, focusing on Postgres
This (currently) small PR adds new `Date` datatypes to Ibis and opens
to road to make them work with Postgres. The main motivation is that I
have some tables with date columns on Redshift, and I'd like them to
work :)    I'm not sure whether this is the right approach or not, so
please comment away!

Author: Alessandro Andrioni <alessandro.andrioni@nubank.com.br>

Closes #874 from andrioni/dates and squashes the following commits:

5c54f64 [Alessandro Andrioni] Add preliminary support for dates, focusing on Postgres