BUG/CLN: Fix predicates on Selections on Joins
The motivation for this PR is twofold    1) fix a bug where filters on
joined data would either fail or give the wrong result, depending on
which version of pandas you're on  2) refactor the
`execute_selection_dataframe` function because it was getting rather
large and unreadable.    The bug in #1136 was occurring due to the
fact that the predicates in the `Selection` were not being evaluated
against the `data` argument  which is the result of the joined data.
This doesn't fail on non-`Join` selections because there's no possible
ambiguity regarding which column to select. The solution is to map
each root table in each predicate to `data`, therefore evaluating the
predicate against the joined data.    Closes #1136.

Author: Phillip Cloud <cpcloud@gmail.com>

Closes #1149 from cpcloud/fix-selection-join-predicates and squashes the following commits:

d76b2a8 [Phillip Cloud] BUG/CLN: Fix predicates on Selections on Joins