BUG: can't mutate and then drop
closes #1296     The issue is that `_maybe_resolve_exprs` was not
ignoring columns that don't exist in a child `Selection` so it would
try to resolve them and throw an `IbisTypeError`. This is okay,
because the expressions have already been resolved in the `Projector`
constructor so trying to project a column that doesn't anywhere would
raise this exception long before getting to `_maybe_resolve_exprs`.

Author: Phillip Cloud <cpcloud@gmail.com>

Closes #1299 from cpcloud/fix-drop and squashes the following commits:

123b156 [Phillip Cloud] If we look for a column that does not exist in the root's selection we cannot fuse so return None
74b2e7f [Phillip Cloud] Cleanup repeated variable use
2b8b34b [Phillip Cloud] Revert
1e31609 [Phillip Cloud] Do something different for drop
de1799c [Phillip Cloud] Test stuff
6bffb3a [Phillip Cloud] Actually exercise the bug
719d0f9 [Phillip Cloud] Add test