ENH: Pandas UD(A)Fs
This PR introduces a way for users to hook their own functions into
the ibis  expression tree for the pandas backend.    Here are some
things to be aware of:    1. Inputs to the function can be any number
of columns (`Series` during     execution) or scalars (Python + numpy
scalar types), Tables (`DataFrame`s)     are not allowed as inputs.
Whether or not we should allow tables as inputs     is an open
discussion.  1. Reductions also take any number of the above types of
inputs, so something     like computing the correlation coefficient is
possible in a group by using     this machinery  1. ~User defined
functions must accept `**kwargs`.~ Functions can opt-in to specific
keyword arguments or all of them with `**kwargs`. The UDF mechanism
decides what to pass in based on the signature of the UDF.  1. Users
must specify the input and output types of the function in the
`@udf`/`@udaf` decorators.    After this is merged, I'd like to
rewrite all of the pandas `execute_node`  definitions using the
UDF/UDAF machinery which would reduce the amount of code  by about a
factor of 2 to 3 for that backend.

Author: Phillip Cloud <cpcloud@gmail.com>

Closes #1277 from cpcloud/pandas-udf and squashes the following commits:

09903c7a [Phillip Cloud] Fix version again
fa97b725 [Phillip Cloud] Check the version of setuptools version instance
5bdf45c7 [Phillip Cloud] Add UDF documentation
e5426159 [Phillip Cloud] Backend details
ccbb18f3 [Phillip Cloud] Fix compat for PY2
fd0b7138 [Phillip Cloud] Only pass in keyword arguments requested by the user in the UDF/UDAF signature
3d62765a [Phillip Cloud] ENH/EXPT: Add pandas UDF provisional API