ENH: Refactor the pandas backend
This is a major refactor of the pandas backend to support new
functionality as  well as to clean up and simplify the architecture.
1. `execute_first` is removed: This was originally useful for
implementing the  ``WindowOp`` rule, but is made obsolete by allowing
``ibis.client.Client``s to  be data arguments in ``execute_node``
rules.  1. `data_preload` is removed: This was originally use to load
in data from the  pandas backend specifically, and was very much a
special snowflake in the  execute loop. This is also made obsolete by
allowing ``ibis.client.Client``s to  be data arguments in
``execute_node`` rules.  1. ``execute_literal`` was introduced to
avoid the performance overhead of  dispatching a call to
``execute_node`` for the few cases we have of evaluating
``ops.Literal`` nodes.  1. I've introduced a new ``inputs`` property
to the ``Node`` class, which is  the sequence of arguments required to
evaluate an expression without any  additional context.  Currently,
only ``WindowOp`` requires something different  because it needs
access to the column underlying the ``expr`` since you cannot  blindly
evaluate ``expr`` and get the correct result, all other ``Node``
subclasses define ``inputs`` as ``return self.args``.  1. Finally,
I've added support for expressions in the `preceding` and `following`
arguments of `WindowOp`.

Author: Phillip Cloud <cpcloud@gmail.com>

Closes #1441 from cpcloud/pandas-execute-refactor and squashes the following commits:

168e016 [Phillip Cloud] ENH: Refactor the pandas backend