BUG: Fix a bug in TableExpr.drop
This PR fixes a bug where the drop method wasn't using `__getitem__`
to  project.    `__getitem__` calls a function `bind_expr` that binds
a table to the projected  expression.    As a follow-up we should
bring this behavior to projection to avoid issues like  this in the
future.    One issue with adding a call to `ir.bind_expr(self, what)`
to the `projection` method is that it breaks some fusion
optimizations.    I think that as part of the the post 1.0.0 follow-up
work we need to disable and remove all optimizations until we can
restructure the compiler flow to separate optimization from expression
construction.

Author: Phillip Cloud <cpcloud@gmail.com>

Closes #1732 from cpcloud/fix-drop and squashes the following commits:

6ab7520 [Phillip Cloud] Implement fix to drop method
f74a8e0 [Phillip Cloud] Fix test
acf53fa [Phillip Cloud] Move test
ed5f086 [Phillip Cloud] Add comment with old generated code
23629f2 [Phillip Cloud] Expected test
461711f [Phillip Cloud] Simplify test
2401f81 [Phillip Cloud] Add test