BUG: Patch meta for DatetimeTZDtype (#2763)
Currently you can't pass in a `pd.DatetimeTZDtype`  to meta. I have a PR up for this in https://github.com/dask/dask/pull/7627, but we can patch this here until that goes through and we update min dependencies for the dask backend.

This is the only place we're using the output meta in this way to create a scalar (it works fine in other contexts).