refactor(datafusion): align register API across backends
BREAKING CHANGE: `register_csv` and `register_parquet` are removed.
Pass filename to `register` method instead.

This also adds in a general backend test for the `register` API so we
can make sure things are consistent, and fixes a small bug with the
dispatch in polars (and datafusion which I copied) that raises a
recursion error if you pass in garbage.

Update ibis/backends/tests/test_client.py

Co-authored-by: Phillip Cloud <417981+cpcloud@users.noreply.github.com>