refactor: rework registration / file loading
Introduces explicit `read_` methods for various types of file /
in-memory object loading.

Removes the `RegexDispatcher` from the DuckDB, polars, and datafusion backends

Should be backwards compatible with existing register functionality,
plus users can now pass in iterables of CSV or parquet files (duckdb only), and also
make use of the explicit `read_` methods if they choose (say if filenames
are non-standard)

Somehow, Datafusion can't handle a `count` so I've added a gross
conditional to the tests for the moment.

This also removes the last usage of RegexDispatcher, so I've removed
that from `dispatch.py` and also removed the odo license file, which is
now no longer required.