fix(common): fix bug in re_extract
Background
==========
The current definition of `re_extract` needs to be clarified. The documentation
claims that it works as `re.match`, but it works differently:

> when index is zero and there's a match, return the entire string, otherwise
> return the content of the index-th match group.

A Python's Match Object returns the entire match, not the entire string.

Implementation
==============
This PR updates the documentation and the implementation of some of the backends
to actually work as a Match Object, note that some backends were already doing
this: pandas, sqlite, duckdb, postgresql.

In the case of BigQuery the index parameter makes no sense because it returns either
the whole match of the first capturing group that matches.

fixes #6167