docs(css): enable code block copy button, don't select prompt
This adds a copy button to every code block.

Also, unused here, but if we use `pycon` for Python code examples, a
leading `>>>` will not be selected if the user tries to select with the
mouse.  (However, the "copy" button still selects the prompt characters,
can't turn that off apparently).