chore(pytest): fail fast on typos in notimpl, notyet, and never marks
A typo in one of our custom marks would just lead to a test
failure (usually) because a test that we expected to xfail wasn't
xfailed.

This is annoying because a) it's surprising and b) you have to run
possibly the whole test suite to hit it (and then you have to figure out
why).

I added a collection hook to flag any typos in the arguments to our
custom marks, and it spits them all out, along with path and test name,
then aborts the test run altogether.