chore(docs): import GroupedTable from real module
GroupedTable is actually defined in the `ibis.expr.types.groupby`
module, not the `ibis.expr.types.relations` module.
Before, quartodoc was noticably slower to build this page of
documentation than the other pages,
because it looked in the wrong
place and then had to do another
import to actually get to the
right place. Before it took
~10s to build this page, now it
is instantaneous.
Not a big deal, but is nice to
iterate faster when working on
the reference docs.