ci(codespell): separate codespell config into standalone file
I really like `pyproject.toml` and having one place to look.  But
running the full backend test suite because we need to add a word to an
"ignore-list" for codespell is silly.

So, breaking it out into a `.codespellrc` file so that it can be updated
without triggering a full CI run.